% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_helpers.R
\name{encode_ddi_factor}
\alias{encode_ddi_factor}
\title{Encode a factor variable using DDI category values and labels}
\usage{
encode_ddi_factor(x, var_name, year, ddi_nac_vars)
}
\arguments{
\item{x}{Vector of values to be encoded as a factor}

\item{var_name}{Name of the variable to be encoded}

\item{year}{Year of the DDI metadata to use}

\item{ddi_nac_vars}{A data frame containing DDI metadata as returned by
\code{ingest_eevv_ddi}.}
}
\value{
A factor vector with the values encoded using the DDI metadata
}
\description{
This function takes a vector of values and encodes them as a factor using the
category values and labels from DDI metadata. It retrives the labels from
\code{ddi_nac_vars} for the respective \code{var_name} and \code{year}
}
