% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_helpers.R
\name{encode_factor_from_haven_labelled_var}
\alias{encode_factor_from_haven_labelled_var}
\title{Encode Factor from Haven Labelled Variable}
\usage{
encode_factor_from_haven_labelled_var(var_to_encode, haven_labelled_var)
}
\arguments{
\item{var_to_encode}{The variable to encode as a factor.}

\item{haven_labelled_var}{A Haven labelled variable from which to extract
levels and labels.}
}
\value{
A factor variable encoded from the Haven labelled variable.
}
\description{
An old function using to encode factor, but using the labels
available in another variable/dataset encoded as a haven_labelled.
We used this, to encode a non-haven_labelled variable (e.g. from other year)
into a factor using the labels from a haven_labelled variable
}
