% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_labels_only.R
\name{get_var_labels_only}
\alias{get_var_labels_only}
\title{Get only variable name/label pairs}
\usage{
get_var_labels_only(ddi_nac_vars)
}
\arguments{
\item{ddi_nac_vars}{A data frame containing DDI metadata as returned by
\code{ingest_eevv_ddi}.}
}
\value{
named character vector
}
\description{
DDI data ingested into ddi_nac_vars contains variable labels for
every year and every variable. But in the end we want to row-bind
all years and hence, we may only want to have one label per variable.
This function makes just that, by aggregating the data in ddi_nac_vars
to select only one label per variable name and return it as a
named character vector, where the name if the variable name and
the value is the label. Thus, it can be conveniently queried like
\code{ddi_nac_var_labels[varname]} to obtain the label associated to \code{varname}.
}
