% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologate_vars.R
\name{homologate_vars}
\alias{homologate_vars}
\title{Homologate variables across years}
\usage{
homologate_vars(nac_raw, ddi_nac_vars, ...)
}
\arguments{
\item{nac_raw}{a dataframe with nacimientos data, including
columns year and sav_data, that containes nested data frame for each year
as returned by \code{ingest_eevv_nac}}

\item{ddi_nac_vars}{A data frame containing information about variables from
Nacimientos in DDI files as returned by \code{ingest_eevv_ddi}}
}
\value{
a data frame with a similar structure as the input nac_raw
(nested data frames in the sav_data variable), but one that makes sure
the variables are all comparable so that they can be row-binded
}
\description{
This will be a monster function!, Sorry about that!
Basically it contains a -hopefully short- code snippet for every
variable in the nacimientos dataset. The goal is to make sure
the same variable is comparable across years. This requires
sometimes recoding levels/values/labels because variables have
changed over time (e.g. new categories introduced, new variables,
or removed variables). So for each variable we will decide what to do
individually, using the data summary in a previous target inspect_vars_info
}
\seealso{
\code{ingest_eevv_nac}, \code{ingest_eevv_ddi}
}
