% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_eevv_ddi.R
\name{ingest_eevv_ddi}
\alias{ingest_eevv_ddi}
\title{Ingest EEVV Nacimientos DDI information}
\usage{
ingest_eevv_ddi()
}
\value{
A data frame containing information about variables from
Nacimientos in DDI files. The data frame includes:
\itemize{
\item \code{xml}: Character vector indicating the source XML file for each variable.
\item \code{id}: Character vector containing the variable ID extracted from the DDI document.
\item \code{files}: Character vector with the data file ID associated in the DDI.
\item \code{uri}: Character vector containing the URI from the DDI document
(potentially useful for further exploration).
\item \code{year}: Numeric year extracted from the URI or variable name
\item \code{name}: Character vector containing the variable name in lowercase
name and year are unique and hence the key to locate the metadata
\item \code{intrvl}: some sort of data type
\item \code{labl}: Character vector with the label associated to the variable
\item \code{catgry}: Nested dataframe with the variable's categories.
the dataframe has columns catValu and labl
}
}
\description{
This function retrieves information about variables related to
births ("Nacimientos") from DANE's EEVV, using the DDI XML files.
Heavy lifting via \code{daner::list_ddi_files} and \code{daner::get_ddi_vars}
}
\details{
But why doing this?, well, DANE's data do not always come with all
the variables encoding information. Although some of the .sav or .dta
files do have labels and values/categories, in several cases they do not
have that. Thus you would have to add them manually, or use some of the
metadata from other years. But you can also extract that metadata
from the DDI. In particular, we want the variable labels and the
categories encoding (value/label pairs).
}
\seealso{
\code{daner::list_ddi_files}, \code{daner::get_ddi_vars}
}
