% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_eevv_nac.R
\name{ingest_eevv_nac}
\alias{ingest_eevv_nac}
\title{Ingest EEVV Nacimientos data}
\usage{
ingest_eevv_nac()
}
\value{
A data frame containing the births data. The data frame includes:
\itemize{
\item \code{bundle}: A character string identifying the file/year the data originated from.
\item \code{sav_file}: Name of the sav file read (set daner::ingest_zipped_sav).
\item \code{size}: Size of the sav file read (set daner::ingest_zipped_sav).
\item \code{date}: Date of the sav file read (set daner::ingest_zipped_sav).
\item \code{year}: Numeric year extracted from the file name.
\item \code{sav_data}: Nested data frame, each representing data from one year.
Each data frame includes the data as is, meaning, it is just the
result of reading the data file via daner::ingest_zipped_sav, which
ultimately reads that using haven::read_sav. So, no further processing
is donde here, except for i) cleaning variable names via janitor::clean_names
and ii) adding a rowid within each dataframe, to make it easier to identify
individual observations and spot/debug errors down the line
}
}
\description{
This function ingests "Nacimientos" (births) data from DANE's EEVV
Data files have been downloaded and stored in the data/EEVV folder.
Yet, DANE does not allow you to make the data available to others
in a public repository. So you probably have to download it yourself.
It identifies relevant files, extracts data using \code{ingest_zipped_sav},
and performs basic cleaning and processing.
}
\seealso{
\code{daner::ingest_zipped_sav}
}
