% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_helpers.R
\name{set_var_label}
\alias{set_var_label}
\title{Set Variable Label}
\usage{
set_var_label(x, var_name, year, ddi_nac_vars)
}
\arguments{
\item{x}{The variable to set the label for.}

\item{var_name}{The name of the variable.}

\item{year}{The year associated with the variable.}

\item{ddi_nac_vars}{A data frame containing DDI metadata as returned by
\code{ingest_eevv_ddi}.}
}
\value{
The variable with the label attribute set.
}
\description{
This function sets the label attribute of a variable, retriving
the labels from the \code{ddi_nac_vars} using \code{var_name} and \code{year}
}
\seealso{
\code{ingest_eevv_ddi}
}
